Sprint 1.28 (3/n): ECO_MAX_DEPOSITS 32 -> 64, con bump de formato
Tercer incremento. El mapa ya usaba 22 de 32 depositos: con 10 libres no caben
las granjas de dos jugadores.

DELEGADO Y VERIFICADO. Contrato PEQUENO (1,5 KB, una sola cosa) a
deepseek-v4-flash con el agente `delegado`. Funciono, y el contraste con el
intento anterior es la leccion del dia: el mismo modelo, con un contrato de
7 KB y cinco partes, se agoto explorando el repositorio y no escribio nada. No
era el modelo: era el contrato.

Encontro 3 desplazamientos de 32 bits que con 64 depositos serian
comportamiento INDEFINIDO —dos en economy.hpp y uno en step.hpp, todos
`uint32_t{1} << i` con i hasta 63— y los corrigio a `uint64_t{1}`. Ese es
exactamente el fallo clasico de este cambio y el contrato se lo senalaba.

Y ENCONTRO UN FALLO MIO, que es lo mas util del sprint. Mis dos commits
anteriores ampliaron EcoDeposit de 4 a 9 campos y rompieron la compilacion de
tres pruebas que usan inicializacion agregada, porque -Wmissing-field-
initializers con -Werror lo rechaza. Mi `ctest` daba 49/49 porque solo habia
recompilado UN objetivo y los demas .o eran viejos. Verificado a mano: HEAD
limpio da 36 errores de compilacion. Es la misma clase de error del binario
rancio que ya me habia mordido antes, y esta vez lo caza el delegado.

BUMP DE VERSIONES, decision mia y no del contrato: SAVE_FORMAT_VERSION 17->18 y
CHECKSUM_ALGO_VERSION 12->13. El agente lo dejo abierto correctamente en vez de
decidirlo por su cuenta. Hacen falta porque `serialize.hpp` escribe
ECO_MAX_DEPOSITS entradas: un guardado viejo tiene 32 y el lector nuevo espera
64, asi que sin bump se malinterpreta EN SILENCIO. Precedente exacto en el 1.9C
con RESOURCE_COUNT.

FALSA ALARMA DE RENDIMIENTO, dicha para que no cunda: el banco marco Step() en
2886 us y checksum en 792, ambos fuera de presupuesto. Era artefacto de MI
build limpio, configurado sin CMAKE_BUILD_TYPE. Con optimizacion: 551 y 128 us,
MEJOR que las referencias y dentro de presupuesto. Duplicar los depositos no
costo nada. Comparar builds de distinto tipo produce regresiones imaginarias.

Baselines re-registrados: G1, G3, G4, skirmish, eco y apertura. Los end_tick NO
se mueven —1227, 1108 y 10473—: cambia lo que se hashea, no como se juega.

Suite 49/49 desde build limpio y optimizado.

+++ addons/chunsa_sim/core/include/chunsa/step.hpp
@@ -1624,7 +1624,7 @@ inline bool is_complete_owned_building(const GameState& g, uint32_t entity,
 // depósitos: evita multiplicar el barrido por MAX_EMITTERS. Todo es ascendente,
 // entero y fijo: cero heap/STL dentro de Step().
 inline void allied_auto_gather_deposit_masks(
-        const GameState& g, uint32_t out[MAX_EMITTERS]) noexcept {
+        const GameState& g, uint64_t out[MAX_EMITTERS]) noexcept {
     for (uint32_t owner = 0; owner < MAX_EMITTERS; ++owner) out[owner] = 0u;
     const uint64_t radius_sq =
         static_cast<uint64_t>(ECO_AUTO_GATHER_RADIUS_RAW)
@@ -1639,17 +1639,17 @@ inline void allied_auto_gather_deposit_masks(
             FatalReason local_fatal = FatalReason::NONE;
             const Vec2Fx deposit{Fx{g.deposits[d].x_raw}, Fx{g.deposits[d].y_raw}};
             if (dist_sq_raw(deposit, building, local_fatal) <= radius_sq) {
-                out[owner] |= (uint32_t{1} << d);
+                out[owner] |= (uint64_t{1} << d);
             }
         }
     }
 }
 
 // Conveniencia de consulta aislada para pruebas del predicado dinámico.
-inline uint32_t allied_auto_gather_deposit_mask(const GameState& g,
+inline uint64_t allied_auto_gather_deposit_mask(const GameState& g,
                                                 uint8_t owner) noexcept {
     if (owner >= MAX_EMITTERS) return 0u;
-    uint32_t eligible[MAX_EMITTERS] = {};
+    uint64_t eligible[MAX_EMITTERS] = {};
     allied_auto_gather_deposit_masks(g, eligible);
     return eligible[owner];
 }
@@ -1714,7 +1714,7 @@ inline void economy_system(GameState& g) noexcept {
     // entrega a cada ciudadano. La máscara solo interviene si SEEK necesita
     // auto-reasignar: un GATHER explícito ya válido puede seguir fuera de zona
     // hasta que su depósito se agote (§23.3).
-    uint32_t auto_gather_eligible[MAX_EMITTERS] = {};
+    uint64_t auto_gather_eligible[MAX_EMITTERS] = {};
     allied_auto_gather_deposit_masks(g, auto_gather_eligible);
 
     for (uint32_t i = 0; i < t.capacity; ++i) {
